requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Session==0.5.0
SQLAlchemy==2.0.48
Werkzeug==3.0.0
gtts==2.4.0
googletrans==4.0.0rc1
python-dotenv==1.0.0
twilio==8.10.0
psycopg[binary]==3.3.3
qrcode==8.0
openpyxl==3.1.2
pytz
